Despues e perder medio cerebro y medio dia pude arreglar unos bugs de la planificacion y ya todas las pruebas del checkpoint funcionan como deberian

+++ planificador/planificador.h
@@ -45,7 +45,7 @@ typedef struct esi_information {
 
 	int esi_id;
 	int esi_numeric_name;
-	double last_estimated_burst;
+	float last_estimated_burst;
 	int last_real_burst;
 	int waited_bursts;
 } esi_information;
